# src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Products
  license:
    name: MIT
paths:
  /product:
    get:
      summary: 获得所有商品信息
      operationId: listProducts
      tags:
        - product
      responses:
        '200':
          description: A paged array of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /product/{productId}:
    get:
      summary: 查看id为productId的商品详情
      operationId: showProductById
      tags:
        - product
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /categories/{categoryName}:
    get:
      summary: 获取类型为categoryName的商品列表
      operationId: showProductsByCategory
      tags:
        - product
      parameters:
        - name: categoryName
          in: path
          required: true
          description: 商品类别名称
          schema:
            type: string
      responses:
        '200':
          description: 成功获取该类别的商品列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        '404':
          description: 该类别不存在或者没有商品属于该类别
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /order:
    get:
      summary: 获取所有订单信息
      operationId: listOrders
      tags:
        - order
      responses:
        '200':
          description: A paged array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: 提交一个订单
      operationId: createOrder
      tags:
        - order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        '200':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /order/{orderId}:
    get:
      summary: 获取订单id为orderId的订单详情
      operationId: showOrderById
      tags:
        - order
      parameters:
        - name: orderId
          in: path
          required: true
          description: The id of the order to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search/{keyword}:
    get:
      summary: 搜索关键字为keyword的商品
      operationId: searchProductsByKeyword
      tags:
        - product
      parameters:
        - name: keyword
          in: path
          required: true
          description: 搜索关键字
          schema:
            type: string
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
        quantity:
          type: integer
        name:
          type: string
        stock:
          type: integer
        img:
          type: string
    Products:
      type: array
      items:
        $ref: "#/components/schemas/Product"
    Order:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              quantity:
                type: integer
      required:
        - items
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string